requirements.txt
streamlit==1.31.0
python-dotenv
supabase
resend
pandas
pyjwt
